Makefile: disable TDLS
NOTE: Since the driver was updated the description below might not be
correct anymore.

This was also disabled for the Citation project. Apparently the Broadcom
driver currently has issues with handling TDLS leading to crashes. Since
we currently don't need it, we disable it because devices might want to
initiate a TDLS connection and crash the driver without our
intervention.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>