Makefile: disable ARP offloading
NOTE: Since the driver was updated the description below might not be
correct anymore.

If ARP offloading is active in AP mode, the ARP responenses from the AP
will have the wrong MAC address (addr[5] + 2).

Looking at `wl_cfg80211_start_ap()` inside `wl_cfg80211.c` the driver is
supposed to disable ARP offloading when the AP is started. However there
seems to be a wrong check for `DHD_FLAG_STA_MODE`, this check is not present
in e.g. the driver for the Cit project. In theory removing the check for the
flag when starting and stopping the AP would work, but only with a single
AP. For the use-case of multiple APs this will break because the code will
not check if any other APs are active and unconditionally re-enable ARP
offloading when an AP is stopped.

For that reason let's just disable ARP offloading completely, when the need
ever arrives we know how to fix it.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>