dhd: fix handling when setting MAC addresses from user-space
If the MAC address is set from user-space, the driver will also write said
address to the `cur_etheraddr` variable inside the chip. However if the
interface is brought down and up again, it seems like this region inside
the WiFi firmware is reset and thus the default (OTP) MAC address is in
the `cur_etheraddr` variable. To circumvent this, we track a previously
written MAC address to `cur_etheraddr` (on wlan0) and if the interface is
brought up again we store this backed-up address back into the running
firmware.

In addition to that setting the MAC address on a downed interface was
previously not working, this is now also fixed by utilizing the mechanism
of writing back a requested MAC address to the firmware on bringing the
interface up.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>