wl_cfg80211: always compile in wl_cfg80211_reg_notifier
CONFIG_CFG80211_INTERNAL_REGDB is long gone (since kernel v4.15) and is
definitely NOT what one would want to use as it bypasses the regdb given
by crda with the regdb from the kernel git repo which is outdated and
one would not want to maintain it anyway.

The reg_notifier callback in the net wireless subsystem is not
compiled in depending on a Kconfig option so it should be safe to always
compile Ampak's implementation of the callback.

This allows the driver to react to changes done to the regulatory domain
when using `iw reg set`.

Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>
[correctly applied on new driver release]
Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>