[HACK] respect regulatory domain as requested by the kernel or APs [STREAMEIGHT-4585]
When the interface is brought down and up, a hardcoded regdom (CN) was
set as default. This means that any call to `iw reg set` while the
interface is not up will not be respected.

Since the kernel does not call the reg_notifier callback when `iw reg
set` is run with the same value as currently returned by `iw reg get`
(or via association to an AP advertising its regdom), a simple userspace
hack cannot work (especially since there's no udev event for when an
interface is brought up or down).

Also, the reg_notifier callback was only registered once the interface
is brought up which meant that the "at-probe" reg_notifier call by the
kernel to give the driver the current regdom wasn't done.

This makes sure the reg_notifier callback is registered before
wiphy_register is called which is the signal for the kernel to call the
reg_notifier callback (if it exists) with the current regdom.

Since the reg_notifier callback can now be executed before the firmware
is flashed (the firmware is flashed only when the interface is brought
up, which can be done much later after the probe of the driver) which
means the main objective of reg_notifier callback to configure the
firmware correctly cannot be done in all cases (e.g. when the interface
is down of hasn't been brought up yet).

Therefore, a copy of the requested alpha2-code regdom is kept so that in
case the interface is brought down later (in which case the "default"
regdom is used when brought up again), it can be applied by the code
taking care of firmware configuration after flashing. This also means
that the last reg_notifier call while the interface is down will be kept
for use after the interface is brought up again later.

Marking this as a hack because I'd rather see something fixing the need
of (re)flashing the firmware every time the interface is brought up.

Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>
[return ret only if kernel version is < 3.10.11, otherwise just return]
Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>
[added logic for clm-blob based chips such as AP6275S]
Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>
[adjusted for new driver version]
Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>