dhd_linux: support 00 regdom on modules using clm blob when set with iw reg [STREAMEIGHT-4659]
When 00 regdom is requested while the interface is up by calling iw reg
set 00, the regdom switching used to fail because 00 regdom is not
defined in the clm (assumed acronym for *c*ountry *l*ist *m*ap) blob.

For many Broadcom modules, the 00 (World) regulatory domain is
internally named "XZ".

The code was copied from some parts in the same function which are
surrounded by a check on DHD_BLOB_EXISTENCE_CHECK and absence of
CUSTOM_COUNTRY_CODE. The issue is that BCM4359 does not support clm blob
and relies on hardcoded country table that is compiled in when
CUSTOM_COUNTRY_CODE is defined.

Moreover, a check on 00 iso_code is performed so that unknown AND World
regdom defaults to the internal representation of XZ.

Finally, one knows the regulatory domain is correct at runtime when the
following messages appear on console:
[dhd] dhd_get_customized_country_code: ccode change to XZ
[dhd] dhd_conf_set_country : set country XZ, revision -1
[dhd] dhd_conf_set_country : Country code: XZ (XZ/0)

This is required in order to support AP6275S (based on BCM43752) WiFi
module from Ampak since it uses a clm blob as opposed to AP6398S
(BCM4359).

Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>
[adjusted for new driver version]
Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>