dhd_config: support 00 regdom on modules using clm blob when set before probing [STREAMEIGHT-4659]
When 00 regdom is requested while the interface is down by calling iw reg
set 00 (or right after boot, by the kernel), the regdom switching used
to fail because the requested regdom was not curated through
dhd_get_customized_country_code which is what's done in the reg_notifier
callback called when executing `iw reg set` or when an AP the DUT
connects to advertises its regdom.

The code was borrowed from wldev_set_country used (indirectly) by the
reg_notifier callback implementation.

Finally, one knows the regulatory domain is correct at runtime when the
following messages appear on console:
[dhd] dhd_get_customized_country_code: ccode change to XZ
[dhd] dhd_conf_set_country : set country XZ, revision -1
[dhd] dhd_conf_set_country : Country code: XZ (XZ/0)

This is required in order to support AP6275S (based on BCM43752) WiFi
module from Ampak.

Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>
[adjusted for new driver version]
Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>