wl_cfg80211: update Linux regdom info for cosmetic purposes
This makes sure that the displayed Linux regdom using `iw reg get` at least
shows the correct country in all cases. The actual regdom is self-managed
by the WiFi chip.

Unforuntately the reported bands in `iw reg get` are just some standard
dummy bands in the default brcm_regdom struct. The proper implementation
would be to read them back form the WiFi chip, but that is not implemented.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>