report carrier state properly [STREAMEIGHT-5258]
Linux has administrative and operational state for an interface (net_device).
This patch adds the ability for the driver to report operational state
(operstate) properly via the netif_carrier_on and netif_carrier_off
calls. These two calls map to __LINK_STATE_NOCARRIER which in turn map
to !IFF_LOWER_UP.

Without calling this function the operstate gets stuck in either
operstate 0 or 6. Both of which signal to the kernel that the interface
is up and running which is not true. This can sometimes cause issues when
wpa_supplicant is started. The kernel initially thinks the driver is
running and reports this to userspace. Some event later corrects this to
operstate 5 (dormant) but it is possible to read the interface as
running before this happens. This can lead to problems when scripts
watching "ip link ls wlan0" get confused by this.

More documentation: Documentation/networking/operstates.txt

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>
[adjusted for new driver version]
Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>